requirements.txt
gradio[mcp]
agno
python-dotenv
google-genai
openai